refactor(model_loader): split load_model into an AITER-free loading core
`load_model` was a 677-line function with ten nested closures inside a
1028-line module, and `loader.py` imports AITER at module level. The unit test
gate has no AITER build, so none of that logic could be covered by a test —
which is how a batched-staging bug that breaks Qwen3.5 BF16 loading reached
main unnoticed.

Move the checkpoint -> parameter logic out into modules that import nothing but
torch, so they can be driven directly from a CPU-only test:

  atom/model_ops/fused_moe/expert_layout.py  expert slot layout: which local
      slots a rank owns (`determine_expert_map`, `count_local_base_experts`)
      and which bytes of a slot one (expert, shard) arrival writes
      (`expert_shard_dim`, `expert_region`)
  atom/model_loader/weight_names.py          `WeightsMapper` plus the
      checkpoint-name helpers (`have_shared_expert`, `shared_expert_prefixes`,
      `extract_expert_target_and_id`)
  atom/model_loader/loading_core.py          `load_weights_into_model`, the
      loading loop itself

`loader.py` keeps the public API and becomes the place where the AITER-shaped
pieces are bound: the TP group (`is_rank0`), quant-config-driven shared-expert
fusion (`fuse_shared_expert`), the safetensors iterator (`weights_iterator`,
also the seam a test needs to control shard read order) and
`default_weight_loader`. Post-load weight processing stays there too.

`expert_region` is new; it is the single definition of the region one arrival
owns, so the staging copy and the staging flush cannot drift apart. Nothing
calls it yet.

Pure refactor: the moved loop body is byte-identical apart from those four
injection points and four mechanical ruff fixes (`getattr(p, "weight_loader")`
-> `p.weight_loader`, `dict.keys()` iteration, tuple `endswith`). Verified by
diffing the extracted range against the original. `moe.py` also picks up the
formatter's pyupgrade pass over the whole file.

Suite unchanged at 759 passed / 6 skipped / 0 failed.
